requirements.txt
pandas>=1.4.0
numpy>=1.22.0
PyYAML>=6.0
